+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This project follows semantic versioning.
 
+### unreleased
+- [added] Integer square root operator `SquareRoot` with alias `Sqrt`.
+
 ### 1.10.0 (2018-03-11)
 - [added] The `PowerOfTwo` marker trait.
 - [added] Associated constants for `Bit`, `Unsigned`, and `Integer`.